# Conflicts:
#	WooCommerce/src/main/kotlin/com/woocommerce/android/WooCommerce.kt
#	WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/CardReaderSettingsFragment.kt
#	build.gradle

+++ WooCommerce/src/main/kotlin/com/woocommerce/android/model/Product.kt
@@ -52,7 +52,7 @@ data class Product(
     val regularPrice: BigDecimal?,
     val taxClass: String,
     val isStockManaged: Boolean,
-    val stockQuantity: Int,
+    val stockQuantity: Double,
     val sku: String,
     val shippingClass: String,
     val shippingClassId: Long,
